atualizando metodos

+++ 2028492_2070405/copiaMatriz.c
@@ -7,10 +7,10 @@ Julio Cesar Rogacheski
 */
 
 #include "libsTrabalho.h"
-double **copiaMatriz(int **mat, int lin, int col)
+int **copiaMatriz(int **mat, int lin, int col)
 {
   int i, j;
-  double **matcopia = alocaMatriz(lin, col);
+  int **matcopia = alocaMatriz(lin, col);
   for (i = 0; i < lin; i++)
   {
     for (j = 0; j < col; j++)
